Allow click action to click non-focusable elements
This assertion was initially to provide some basic a11y checking but it turns
out that an element can be considered interactive without being focusable if it
has one of the widget roles.

Supporting roles will be a slippery slope for making other actions behave
accessibly, such as the `select` action working with custom listbox
elements. For now we can remove this focusable check and maybe in the future
reconsider a11y.