Add auto-defined assertions for normal computed properties
In most cases, user-defined computed properties return simple values. Instead of
having to define assertions for computed getter properties, they can be
auto-defined based on the property name and a defualt matcher.

For user-defined computed getter properties that do not return simple values, an
assertion is still auto-defined, but may not be as useful as manually defining
custom assertions.