Fix incorrect parent class when extending custom interactors
The `from` method sometimes assumed that the bound class was not always an
interactor. Since it is a static interactor method, the bound class should
always be an interactor.

The decorator sometimes bound a non-interactor to the `from` method. Now that
`from` correctly assumes it is _always_ bound to an interactor, the decorator
needs to ensure that it _always_ binds an interactor.