Fix $element argument being eagerly computed
When the `$element` argument is given to assertions, and by extension the `when`
method, it is eagerly computed during `run` which causes an immediate error when
the element is not found. Lazily computing the argument allows assertions and
the `when` method to properly rerun until `$element` does not throw an error.