Fix negated assertions resolving when the element cannot be found
The `not` property sets the expectation for the following assertion to fail;
when an "unable to find element" error occurs, the assertion _technically_
fails, which satisfies the `not` expectation.

Some errors should always bubble through the `not` expectation. These errors can
be created with the new `error` util which sets an internal meta value on the
error to tell the validator to always raise the error.